+++ engine/platform/win32/win_snd.cpp
@@ -889,8 +889,8 @@ static qboolean SNDDMA_InitDS( void )
 
 	memset( &format, 0, sizeof( format ) );
 	format.wFormatTag = WAVE_FORMAT_PCM;
-	format.nChannels = dma.channels;
-	format.wBitsPerSample = dma.samplebits;
+	format.nChannels = (WORD)( dma.channels );
+	format.wBitsPerSample = (WORD)( dma.samplebits );
 	format.nSamplesPerSec = dma.speed;
 	format.nBlockAlign = format.nChannels * format.wBitsPerSample / 8;
 	format.cbSize = 0;
