+++ engine/renderervk/tr_model.cpp
@@ -461,7 +461,7 @@ static qboolean R_LoadMD3( model_t *mod, int lod, void *buffer, int fileSize, co
 	for ( i = 0 ; i < hdr->numSurfaces; i++) {
 
 		// how many bytes are remaining before end of file
-		bytesToEnd = size - ((byte*)surf - (byte*)hdr);
+		bytesToEnd = (uint32_t)( size - ((byte*)surf - (byte*)hdr) );
 
 		if ( bytesToEnd < sizeof(*surf)) {
 			ri.Printf( PRINT_WARNING, "%s: %s has corrupted surface header\n", __func__, mod_name );
@@ -1042,7 +1042,7 @@ static md3Tag_t *R_GetAnimTag( mdrHeader_t *mod, int framenum, const char *tagNa
 
 			// uncompressed model...
 			//
-			frameSize = (intptr_t)( &((mdrFrame_t *)0)->bones[ mod->numBones ] );
+			frameSize = (int)( (intptr_t)( &((mdrFrame_t *)0)->bones[ mod->numBones ] ) );
 			frame = (mdrFrame_t *)((byte *)mod + mod->ofsFrames + framenum * frameSize );
 
 			for (j = 0; j < 3; j++)
