+++ game/cgame/cg_ents.cpp
@@ -215,7 +215,7 @@ static void CG_Speaker( centity_t *cent ) {
 
 	//	ent->s.frame = ent->wait * 10;
 	//	ent->s.clientNum = ent->random * 10;
-	cent->miscTime = cg.time + cent->currentState.frame * 100 + cent->currentState.clientNum * 100 * crandom();
+	cent->miscTime = (int)( cg.time + cent->currentState.frame * 100 + cent->currentState.clientNum * 100 * crandom() );
 }
 
 /*
