+++ game/ui/ui_cdkey.cpp
@@ -151,7 +151,7 @@ static void UI_CDKeyMenu_DrawKey( void *self ) {
 
 	x = 320 - 8 * BIGCHAR_WIDTH;
 	y = 240 - BIGCHAR_HEIGHT / 2;
-	UI_FillRect( (float)( x ), (float)( y ), 16 * BIGCHAR_WIDTH, BIGCHAR_HEIGHT, listbar_color );
+	UI_FillRect( (float)( x ), (float)( y ), (float)( 16 * BIGCHAR_WIDTH ), (float)( BIGCHAR_HEIGHT ), listbar_color );
 	UI_DrawString( x, y, f->field.buffer, style, color );
 
 	// draw cursor if we have focus
